add FieldTypeResolver support
