+++ src/main/java/com/alibaba/fastjson/parser/deserializer/FieldTypeResolver.java
@@ -1,0 +1,7 @@
+package com.alibaba.fastjson.parser.deserializer;
+
+import java.lang.reflect.Type;
+
+public interface FieldTypeResolver extends ParseProcess {
+    Type resolve(Object object, String fieldName);
+}
